+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/KafkaTestUtils.java
@@ -5,6 +5,7 @@ package io.confluent.csid.utils;
  */
 
 import io.confluent.parallelconsumer.AbstractParallelEoSStreamProcessorTestBase;
+import io.confluent.parallelconsumer.PollContext;
 import io.confluent.parallelconsumer.offsets.OffsetMapCodecManager;
 import io.confluent.parallelconsumer.state.WorkContainer;
 import io.confluent.parallelconsumer.state.WorkManager;
@@ -12,6 +13,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.MockConsumer;
 import org.apache.kafka.clients.consumer.OffsetAndMetadata;
@@ -33,6 +35,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class KafkaTestUtils {
 
     private final String INPUT_TOPIC;
+
     private final String CONSUMER_GROUP_ID;
 
     @Getter
@@ -284,5 +287,33 @@ public class KafkaTestUtils {
         var consumerRecords = generateRecords(i);
         send(consumerSpy, consumerRecords);
         return consumerRecords;
+    }
+
+    /**
+     * Checks that the ordering of the results is the same as the ordering of the input records
+     */
+    // todo move to specific assertion utils class, along with other legacy assertion utils?
+    public static <T> void checkExactOrdering(Map<String, Queue<PollContext<String, String>>> results,
+                                              HashMap<Integer, List<T>> originalRecords) {
+        originalRecords.entrySet().forEach(entry -> {
+            var originalRecordList = entry.getValue();
+            var originalKey = entry.getKey();
+            var sequence = results.get(originalKey.toString());
+            assertThat(sequence).hasSameSizeAs(originalRecordList);
+            assertThat(sequence.size()).describedAs("Sanity: is same size as original list").isEqualTo(originalRecordList.size());
+            log.debug("Key {} has same size of record as original - {}", originalKey, sequence.size());
+            // check the integer sequence of PollContext value is linear and is without gaps
+            var last = sequence.poll();
+            PollContext<String, String> next = null;
+            while (!sequence.isEmpty()) {
+                next = sequence.poll();
+                var thisValue = Integer.parseInt(org.apache.commons.lang3.StringUtils.substringBefore(next.value(), ","));
+                var lastValue = Integer.parseInt(StringUtils.substringBefore(last.value(), ",")) + 1;
+                assertThat(thisValue).isEqualTo(lastValue);
+                last = next;
+            }
+            log.debug("Key {} a an exactly sequential series of values, ending in {} (starts at zero)", originalKey, next.value());
+
+        });
     }
 }
